requirements.txt
flask
pywifi 
qrcode[pil] 
pillow 
requests 
comtypes
